Move sensitivity test functions
