--- emukit/sensitivity/test_functions/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-from .ishigami import Ishigami
