+++ emukit/test_functions/sensitivity/__init__.py
@@ -1,0 +1,1 @@
+from .ishigami import Ishigami
