+++ tests/emukit/sensitivity/test_emukit_sensitivity.py
@@ -4,7 +4,7 @@ import random
 
 from emukit.core import ContinuousParameter, ParameterSpace
 from emukit.sensitivity.monte_carlo import MonteCarloSensitivity
-from emukit.sensitivity.test_functions import Ishigami
+from emukit.test_functions.sensitivity import Ishigami
 
 
 def test_montecarlo_sensitivity():
